+++ src/locales/en.js
@@ -115,4 +115,10 @@ export default {
       comingSoon: '(Coming Soon)',
     },
   },
+  apiProduct: {
+    options: {
+      headline: 'Select your preferred variant',
+      pleaseSelect: 'Pick a {{option_name}}',
+    },
+  },
 };