+++ src/sdk.js
@@ -215,10 +215,10 @@ class StorefrontAPI {
   /**
    * Receives data of a specific product.
    * @param {String} productCode The target product code.
-   * @param {Array} fields Data fields that should be fetched.
+   * @param {Array} [fields=[]] Data fields that should be fetched.
    * @returns {Object}
    */
-  async getProduct(productCode, fields) {
+  async getProduct(productCode, fields = []) {
     return this._genericRequest({
       endpoint: `/v1/products/${encodeURIComponent(productCode)}`,
       query: {