+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # retail.red Storefront Library Changelog
 
+## v3.0.0
+### Breaking changes
+- Reworked big parts of HTML markup and CSS
+- Removed default value for --rr-color-text custom property, so that tha page text color is applied to text within the modal
+
+### Changed
+- Set default value for `--rr-color-text` custom property to `inherit`, so that by default the primary page text color is used for texts within the modal
+### Added
+- Added option `useApiProduct` to fetch product data via the Storefront API. Required properties within the product configuration are `code` and `quantity`
+- Introduced optional `--rr-color-secondary` and `--rr-color-secondary-contrast` custom properties to enable special color styling for search buttons and form inputs
+
 ## v2.1.0
 ### Changes
 - Improved accessibility